Refactor for Testing
Modified the TPM fact to make things easier to test.

Provided one working example in the tpm_spec.rb file for testing the
various methods.

The rest should be pretty cut and dry.
